refactor: move `EVMLegacyFilterExec` to separate file
