+++ crates/proof-of-sql/src/sql/evm_proof_plan/plans/mod.rs
@@ -7,8 +7,8 @@ use crate::{
     sql::{
         proof_exprs::{AliasedDynProofExpr, ColumnExpr, TableExpr},
         proof_plans::{
-            AggregateExec, DynProofPlan, FilterExec, GroupByExec, LegacyFilterExec, ProjectionExec,
-            SliceExec, SortMergeJoinExec, UnionExec,
+            AggregateExec, DynProofPlan, FilterExec, GroupByExec, ProjectionExec, SliceExec,
+            SortMergeJoinExec, UnionExec,
         },
     },
 };
@@ -21,6 +21,9 @@ pub(super) use evm_empty_exec::EVMEmptyExec;
 
 mod evm_table_exec;
 pub(super) use evm_table_exec::EVMTableExec;
+
+mod evm_legacy_filter_exec;
+pub(super) use evm_legacy_filter_exec::EVMLegacyFilterExec;
 
 mod conversion_utils;
 pub(super) use conversion_utils::{
@@ -140,64 +143,6 @@ impl EVMDynProofPlan {
                 aggregate_exec.try_into_proof_plan(table_refs, column_refs, output_column_names)?,
             )),
         }
-    }
-}
-
-/// Represents a filter execution plan in EVM.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub(crate) struct EVMLegacyFilterExec {
-    table_number: usize,
-    where_clause: EVMDynProofExpr,
-    results: Vec<EVMDynProofExpr>,
-}
-
-impl EVMLegacyFilterExec {
-    /// Try to create a `LegacyFilterExec` from a `proof_plans::LegacyFilterExec`.
-    pub(crate) fn try_from_proof_plan(
-        plan: &LegacyFilterExec,
-        table_refs: &IndexSet<TableRef>,
-        column_refs: &IndexSet<ColumnRef>,
-    ) -> EVMProofPlanResult<Self> {
-        Ok(Self {
-            table_number: table_refs
-                .get_index_of(&plan.table().table_ref)
-                .ok_or(EVMProofPlanError::TableNotFound)?,
-            results: plan
-                .aliased_results()
-                .iter()
-                .map(|result| EVMDynProofExpr::try_from_proof_expr(&result.expr, column_refs))
-                .collect::<Result<_, _>>()?,
-            where_clause: EVMDynProofExpr::try_from_proof_expr(plan.where_clause(), column_refs)?,
-        })
-    }
-
-    pub(crate) fn try_into_proof_plan(
-        &self,
-        table_refs: &IndexSet<TableRef>,
-        column_refs: &IndexSet<ColumnRef>,
-        output_column_names: Option<&IndexSet<String>>,
-    ) -> EVMProofPlanResult<LegacyFilterExec> {
-        let output_column_names =
-            try_unwrap_output_column_names(output_column_names, self.results.len())?;
-        Ok(LegacyFilterExec::new(
-            self.results
-                .iter()
-                .zip(output_column_names.iter())
-                .map(|(expr, name)| {
-                    Ok(AliasedDynProofExpr {
-                        expr: expr.try_into_proof_expr(column_refs)?,
-                        alias: Ident::new(name),
-                    })
-                })
-                .collect::<EVMProofPlanResult<Vec<_>>>()?,
-            TableExpr {
-                table_ref: table_refs
-                    .get_index(self.table_number)
-                    .cloned()
-                    .ok_or(EVMProofPlanError::TableNotFound)?,
-            },
-            self.where_clause.try_into_proof_expr(column_refs)?,
-        ))
     }
 }
 
@@ -728,7 +673,6 @@ mod tests {
             map::indexset,
         },
         sql::{
-            evm_proof_plan::exprs::{EVMColumnExpr, EVMEqualsExpr, EVMLiteralExpr},
             proof::ProofPlan,
             proof_exprs::{
                 AddExpr, AliasedDynProofExpr, ColumnExpr, DynProofExpr, EqualsExpr, LiteralExpr,
@@ -879,75 +823,6 @@ mod tests {
             .unwrap();
 
         assert_eq!(dyn_slice_exec, DynProofPlan::Slice(slice_exec));
-    }
-
-    #[test]
-    fn we_can_put_filter_exec_in_evm() {
-        let table_ref: TableRef = "namespace.table".parse().unwrap();
-        let ident_a = "a".into();
-        let ident_b = "b".into();
-        let alias = "alias".to_string();
-
-        let column_ref_a = ColumnRef::new(table_ref.clone(), ident_a, ColumnType::BigInt);
-        let column_ref_b = ColumnRef::new(table_ref.clone(), ident_b, ColumnType::BigInt);
-
-        let filter_exec = LegacyFilterExec::new(
-            vec![AliasedDynProofExpr {
-                expr: DynProofExpr::Column(ColumnExpr::new(column_ref_b.clone())),
-                alias: Ident::new(alias.clone()),
-            }],
-            TableExpr {
-                table_ref: table_ref.clone(),
-            },
-            DynProofExpr::Equals(
-                EqualsExpr::try_new(
-                    Box::new(DynProofExpr::Column(ColumnExpr::new(column_ref_a.clone()))),
-                    Box::new(DynProofExpr::Literal(LiteralExpr::new(
-                        LiteralValue::BigInt(5),
-                    ))),
-                )
-                .unwrap(),
-            ),
-        );
-
-        let evm_filter_exec = EVMLegacyFilterExec::try_from_proof_plan(
-            &filter_exec,
-            &indexset![table_ref.clone()],
-            &indexset![column_ref_a.clone(), column_ref_b.clone()],
-        )
-        .unwrap();
-
-        let expected_evm_filter_exec = EVMLegacyFilterExec {
-            table_number: 0,
-            where_clause: EVMDynProofExpr::Equals(EVMEqualsExpr::new(
-                EVMDynProofExpr::Column(EVMColumnExpr::new(0)),
-                EVMDynProofExpr::Literal(EVMLiteralExpr(LiteralValue::BigInt(5))),
-            )),
-            results: vec![EVMDynProofExpr::Column(EVMColumnExpr::new(1))],
-        };
-
-        assert_eq!(evm_filter_exec, expected_evm_filter_exec);
-
-        // Roundtrip
-        let roundtripped_filter_exec = EVMLegacyFilterExec::try_into_proof_plan(
-            &evm_filter_exec,
-            &indexset![table_ref.clone()],
-            &indexset![column_ref_a.clone(), column_ref_b.clone()],
-            Some(&indexset![alias]),
-        )
-        .unwrap();
-        assert_eq!(roundtripped_filter_exec, filter_exec);
-
-        assert!(matches!(
-            EVMLegacyFilterExec::try_into_proof_plan(
-                &evm_filter_exec,
-                &indexset![],
-                &indexset![],
-                Some(&indexset![]),
-            )
-            .unwrap_err(),
-            EVMProofPlanError::InvalidOutputColumnName
-        ));
     }
 
     #[test]
